requirements.txt
Django==1.9.1
django-cors-headers==1.1.0
django-countries==3.4.1
django-crispy-forms==1.5.2
django-filter==0.11.0
django-localflavor==1.2
django-model-utils==2.4
djangorestframework==3.3.2
ecdsa==0.13
Fabric==1.10.2
paramiko==1.16.0
pycrypto==2.6.1
wheel==0.24.0
